rename firebase to firestore
